split dataStorageAccess into a different file to allow for jest mocks with circular deps
