+++ extension/src/background/helpers/dataStorage.ts
@@ -1,4 +1,3 @@
-import browser from "webextension-polyfill";
 import semver from "semver";
 
 import {
@@ -21,66 +20,10 @@ import {
   SOROBAN_RPC_URLS,
 } from "@shared/constants/stellar";
 import { DEFAULT_ASSETS_LISTS } from "@shared/constants/soroban/token";
-
-export interface SetItemParams {
-  [key: string]: any;
-}
-
-// https://github.com/mozilla/webextension-polyfill/issues/424
-interface BrowserStorage extends browser.Storage.Static {
-  session: browser.Storage.LocalStorageArea;
-}
-
-const storage = browser.storage as BrowserStorage;
-
-// browser storage uses local storage which stores values on disk and persists data across sessions
-// session storage uses session storage which stores data in memory and clears data after every "session"
-// only use session storage for secrets or sensitive values
-export const browserLocalStorage = storage?.local;
-export const browserSessionStorage = storage?.session;
+import { dataStorageAccess, browserLocalStorage } from "./dataStorageAccess";
 
 // Session Storage Feature Flag - turn on when storage.session is supported
 export const SESSION_STORAGE_ENABLED = true;
-
-export type StorageOption =
-  | typeof browserLocalStorage
-  | typeof browserSessionStorage;
-
-export const dataStorage = (
-  storageApi: StorageOption = browserLocalStorage,
-) => ({
-  getItem: async (key: string) => {
-    // TODO: re-enable defaults by passing an object. The value of the key-value pair will be the default
-
-    const storageResult = await storageApi.get(key);
-
-    return storageResult[key];
-  },
-  setItem: async (setItemParams: SetItemParams) => {
-    await storageApi.set(setItemParams);
-  },
-
-  clear: async () => {
-    await storageApi.clear();
-  },
-  remove: async (keys: string | string[]) => {
-    await storageApi.remove(keys);
-  },
-});
-
-export const dataStorageAccess = (
-  storageApi: StorageOption = browserLocalStorage,
-) => {
-  const store = dataStorage(storageApi);
-  return {
-    getItem: store.getItem,
-    setItem: async (keyId: string, value: any) => {
-      await store.setItem({ [keyId]: value });
-    },
-    clear: () => store.clear(),
-    remove: store.remove,
-  };
-};
 
 export const normalizeMigratedData = async () => {
   const localStore = dataStorageAccess(browserLocalStorage);
@@ -110,7 +53,7 @@ export const normalizeMigratedData = async () => {
 
 // This migration adds a friendbotUrl to testnet and futurenet network details
 export const migrateFriendBotUrlNetworkDetails = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
 
   const networksList: NetworkDetails[] =
     (await localStore.getItem(NETWORKS_LIST_ID)) || DEFAULT_NETWORKS;
@@ -131,7 +74,7 @@ export const migrateFriendBotUrlNetworkDetails = async () => {
 };
 
 export const migrateSorobanRpcUrlNetworkDetails = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
 
   const networksList: NetworkDetails[] =
     (await localStore.getItem(NETWORKS_LIST_ID)) || DEFAULT_NETWORKS;
@@ -152,7 +95,7 @@ export const migrateSorobanRpcUrlNetworkDetails = async () => {
 
 // This migration migrates the storage for custom tokens IDs to be keyed by network
 export const migrateTokenIdList = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const tokenIdsByKey = await localStore.getItem(TOKEN_ID_LIST);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
@@ -169,7 +112,7 @@ export const migrateTokenIdList = async () => {
 };
 
 export const migrateTestnetSorobanRpcUrlNetworkDetails = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
   if (!storageVersion || semver.lt(storageVersion, "2.0.0")) {
@@ -199,7 +142,7 @@ export const migrateTestnetSorobanRpcUrlNetworkDetails = async () => {
 };
 
 export const migrateToAccountSubscriptions = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
   // we only want to run this once per user
@@ -210,7 +153,7 @@ export const migrateToAccountSubscriptions = async () => {
 };
 
 export const migrateMainnetSorobanRpcUrlNetworkDetails = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
   if (!storageVersion || semver.lt(storageVersion, "4.0.0")) {
@@ -240,7 +183,7 @@ export const migrateMainnetSorobanRpcUrlNetworkDetails = async () => {
 };
 
 export const migrateSorobanRpcUrlNetwork = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
   if (!storageVersion || semver.lt(storageVersion, "4.0.1")) {
@@ -263,7 +206,7 @@ export const migrateSorobanRpcUrlNetwork = async () => {
 };
 
 export const resetAccountSubscriptions = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
   if (!storageVersion || semver.eq(storageVersion, "4.0.2")) {
@@ -274,7 +217,7 @@ export const resetAccountSubscriptions = async () => {
 };
 
 export const addAssetsLists = async () => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
   const storageVersion = (await localStore.getItem(STORAGE_VERSION)) as string;
 
   if (!storageVersion || semver.lt(storageVersion, "4.1.0")) {
@@ -310,7 +253,7 @@ export const versionedMigration = async () => {
 
 // Updates storage version
 export const migrateDataStorageVersion = async (version: string) => {
-  const localStore = exports.dataStorageAccess(browserLocalStorage);
+  const localStore = dataStorageAccess(browserLocalStorage);
 
   // This value should be manually updated when a new schema change is made
   await localStore.setItem(STORAGE_VERSION, version);
